requirements.txt
Flask==3.0.3
prisma==0.15.0
PyJWT==2.9.0

